Update patch set 7
Patch Set 7:

Rather than regexing a 2 line file with a very complex regex pattern. I think it's far simpler to check-in the file. It's easier to understand what's happening in the crontab and it's a file we've already decided to manage by adding these regexes anyway.

I mean this:

 # Run system activity accounting tool
 */1 * * * * root /usr/lib64/sa/sa1 1 1
 # Generate a daily summary of process accounting at 23:53
 53 23 * * * root /usr/lib64/sa/sa2 -A

Is far easier to know exactly what is in that crontab than having to figure out whatever this is:

 ((.*\/)(10)([ \*]+))(\s+.*)(sa1.*)

Additionally there is no question about what is in the first example whereas the 2nd example requires someone reading this file to spawn a VM just to figure out what the resultant crontab is.

Patch-set: 7