Refactor lf-recommended-tools to install in a single transaction
This patch moves the variables to the appropriate vars folder with
the correct ansible_os_family.  It lists all of the common packages
in main.yml and distro-specific in the respective YAML file.

It also makes the installation happen with a single transaction
by providing a list to the package module.

Change-Id: Ib1d1bda84711de33bc1f9b6d120d26585bb55acc
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>